cpu/arm/stm32-ocm3/clock.c: correct reload value; don't accumulate CLOCK_MICROSECOND_SYSTICK rounding errors
This commit sets the reload value with better accuracy. The difference
between (X + Y - 1) / Y - 1 and simply X / Y can get significant for
low values of X, i.e., if we're running with a very low AHB clock.

Furthermore, we allow clock_delay_usec to use wait of up to half the
STK period, to mitigate the effect of rounding errors in
CLOCK_MICROSECOND_SYSTICK.